Create CUFFT plans for double precision real and complex numbers
Note that, only function-like parameters can be pasted in a macro, and
the parameters are not parsed for macro replacement first, so the
following

  CONCATENATION(ExecR2C, ...)

does not expand properly.  Instead, we need to use two "layers" of
macros to ensure names are replaced before concatentation.  This
method is called "indirectly quoting macro arguments".  See
<http://en.wikipedia.org/wiki/C_preprocessor> for details.