Introduce random forcing
We implement a random forcing that is delta-correlated in time, which
is turned on when the forcing parameters satisfies fi * ki > 0.  For
fi * ki < 0, we fall back to the Kolmogorov forcing.

The random forcing takes the form

  f_w = fi ki cos(k.x + phi) sqrt(N),

where fi is some amplitude, |k| == ki is the energy injection
wavenumber.  Both the direction of k and the phase phi are randomly
chosen.  N measures the number of time steps so the extra scaling
sqrt(N) ensure the energy input is independent of dt.  For simplicity,
we choose N = 1 / dt.