Clean up the code and put in more comments
main.cu is getting big so I split the computation loop to solve.cu.
More comments are put in main.cu to make the code easier to read.
Files that use exit() should include stdlib.h.

I was trying to rename some .cu files to pure C files.  This is
non-trivial because we use types such as uint3 in ihd.h, which are not
standard in C.  Also, .cu is kind of like C++ so we have function
overloading.  Let's stick with .cu for a little bit longer...