Restarting the code produces exactly the same result on Linux
In commit ea37fc..., I fixed the code so that restarting the code on
my laptop (a MacBook Pro) reproduced the exact same result.  The same
feature works on my laptop even though we've made couple changes in
the output files.

Unfortunately, when I test the code on a 64-bit Linux machine
(2.6.18), the restart runs fail to produce the exact same result:

  $ ./ihd -s 10 -t 1
  $ ./iht -s 10 -t 1 -o new- 0005.raw
  $ md5sum 0010.raw new-0010.raw
  8a35380f855e8e02b87fb11fdb9af5d7  0010.raw
  bb279f9ed27e261a44583748de915b45  new-0010.raw

This is annoying.  It turns out, on OS X, the following two lines of C
code produce the same sequence of random number

  srand(1); for(...) { rand(); }
  srand(1); for(...) { seed = rand(); srand(seed); }

But this is not true in Linux, at least on the machine that I tested
on.  Therefore, I need to call srand() inside the main loop (solve.cu
line 23) in order to make the results reproducible.