Compute and plot energy and enstrophy fluxes
We correct couple sign errors in the README file.  The script oned.pro
is improved so it can compute the cumulative spectrum; spec.pro is
updated to use the new output of oned.pro.  A new script, flux.pro, is
written to compute and plot the energy and enstrophy fluxes.